Move Anvil transaction state out of transaction resources
